+++ src/shared/config/axios-instance.ts
@@ -1,0 +1,7 @@
+import { initInstance } from '@shared/lib';
+
+export const BASE_URI = `http://algo.knu-soft.site`;
+
+export const fetchInstance = initInstance({
+	baseURL: BASE_URI,
+});
