+++ src/shared/config/index.ts
